Обновлена логика

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoHibernateImpl.java
@@ -10,11 +10,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class UserDaoHibernateImpl implements UserDao {
-    private SessionFactory sessionFactory;
 
-    public UserDaoHibernateImpl(SessionFactory sessionFactory) {
-        this.sessionFactory = sessionFactory;
-    }
+    private final SessionFactory sessionFactory = new Util().getSessionFactory();
 
     private static final String DROP_TABLE = "DROP TABLE IF EXISTS users";
     private static final String CREATE_TABLE = "CREATE TABLE IF NOT EXISTS users (" +
@@ -25,102 +22,87 @@ public class UserDaoHibernateImpl implements UserDao {
 
     @Override
     public void createUsersTable() {
-        Session session = sessionFactory.openSession();
         Transaction transaction = null;
-
-        try {
+        try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
             session.createSQLQuery(CREATE_TABLE).executeUpdate();
             transaction.commit();
         } catch (Exception e) {
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
-        } finally {
-            session.close();
         }
     }
 
     @Override
     public void dropUsersTable() {
-        Session session = sessionFactory.openSession();
         Transaction transaction = null;
-
-        try {
+        try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
             session.createSQLQuery(DROP_TABLE).executeUpdate();
             transaction.commit();
         } catch (Exception e) {
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
-        } finally {
-            session.close();
         }
     }
 
     @Override
     public void saveUser(String name, String lastName, byte age) {
-        Session session = sessionFactory.openSession();
         Transaction transaction = null;
-
-        try {
+        try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
             User user = new User(name, lastName, age);
             session.save(user);
             System.out.println("User с именем – " + name + " добавлен в базу данных.");
             transaction.commit();
         } catch (Exception e) {
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
-        } finally {
-            session.close();
         }
     }
 
     @Override
     public void removeUserById(long id) {
-        Session session = sessionFactory.openSession();
         Transaction transaction = null;
-
-        try {
+        try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
-            User user = (User) session.get(User.class, id);
-            session.delete(user);
-            System.out.println("User с ID " + id + " удалён.");
+            User user = session.get(User.class, id);
+            if (user != null) {
+                session.delete(user);
+                System.out.println("User с ID " + id + " удалён.");
+            }
             transaction.commit();
         } catch (Exception e) {
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
-        } finally {
-            session.close();
         }
     }
 
     @Override
     public List<User> getAllUsers() {
-        List<User> user = new ArrayList<>();
-        Session session = sessionFactory.openSession();
         Transaction transaction = null;
-
-        try {
+        List<User> users = new ArrayList<>();
+        try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
-            user = session.createQuery("from User", User.class).list();
+            users = session.createQuery("from User", User.class).list();
             transaction.commit();
         } catch (Exception e) {
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
-        } finally {
-            session.close();
         }
-        return user;
+        return users;
     }
 
     @Override
     public void cleanUsersTable() {
-        Session session = sessionFactory.openSession();
         Transaction transaction = null;
-        try {
+        try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
             session.createQuery("delete from User").executeUpdate();
             transaction.commit();
         } catch (Exception e) {
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
-        } finally {
-            session.close();
         }
     }
 }
